refactor: deepen target adaptation and feature installation
