+++ CONTEXT.md
@@ -23,3 +23,7 @@ _Avoid_: 对所有版本通用的猜测、散落的混淆名
 **配置 schema**:
 设置进程与 Apple Music 目标进程共享的配置键、默认值、编码和升级规则；任何进程都不得自行复制这些规则。
 _Avoid_: 配置常量集合、各进程独立读取
+
+**目标适配**:
+AM++ 对一个 Apple Music 版本可供各项增强功能使用的目标能力所作的统一语义描述；各项能力独立报告可用、缺失或歧义。
+_Avoid_: 原始反射句柄、功能内散落的私有成员知识
