+++ app/src/test/java/dev/amenhancer/module/hook/LyricHighlightEventRouterTest.kt
@@ -1,0 +1,54 @@
+package dev.amenhancer.module.hook
+
+import android.view.View
+import org.junit.Assert.assertEquals
+import org.junit.Test
+
+class LyricHighlightEventRouterTest {
+    @Test
+    fun `view model drives the current line while callback hook is unavailable`() {
+        val runtime = RecordingLyricBlurRuntime()
+        val router = LyricHighlightEventRouter(runtime)
+
+        router.onFourArgumentViewModelEvent(lineId = 41, isBackground = false)
+        router.onFourArgumentViewModelEvent(lineId = 42, isBackground = true)
+        router.onFourArgumentViewModelEvent(lineId = 0, isBackground = false)
+        router.onSingleArgumentViewModelEvent(lineId = 43)
+        router.onSingleArgumentViewModelEvent(lineId = -1)
+
+        assertEquals(listOf(41, 43), runtime.fallbackHighlights)
+    }
+
+    @Test
+    fun `installed callback suppresses both view model fallback shapes`() {
+        val runtime = RecordingLyricBlurRuntime()
+        val router = LyricHighlightEventRouter(runtime)
+
+        router.onCallbackInstalled()
+        router.onCallback(setOf(8, 9))
+        router.onFourArgumentViewModelEvent(lineId = 10, isBackground = false)
+        router.onSingleArgumentViewModelEvent(lineId = 11)
+
+        assertEquals(listOf(emptySet(), setOf(8, 9)), runtime.highlightUpdates)
+        assertEquals(emptyList<Int>(), runtime.fallbackHighlights)
+    }
+
+    private class RecordingLyricBlurRuntime : LyricBlurRuntime {
+        val highlightUpdates = mutableListOf<Set<Int>>()
+        val fallbackHighlights = mutableListOf<Int>()
+
+        override fun onSessionChanged(songInfo: Any) = Unit
+
+        override fun onHighlightsChanged(lineIds: Set<Int>) {
+            highlightUpdates += lineIds
+        }
+
+        override fun onFallbackHighlightChanged(lineId: Int) {
+            fallbackHighlights += lineId
+        }
+
+        override fun onLyricsViewCreated(owner: Any, root: View) = Unit
+
+        override fun onLyricsViewDestroyed(owner: Any) = Unit
+    }
+}
