+++ app/src/test/java/dev/amenhancer/module/hook/RightLyricsPaneStructuralRegressionTest.kt
@@ -14,10 +14,10 @@ import org.junit.Test
 class RightLyricsPaneStructuralRegressionTest {
     private val source: String by lazy {
         sequenceOf(
-            File("src/main/java/dev/amenhancer/module/hook/DualPaneFeature.kt"),
-            File("app/src/main/java/dev/amenhancer/module/hook/DualPaneFeature.kt"),
+            File("src/main/java/dev/amenhancer/module/hook/AppleMusicDualPaneTarget.kt"),
+            File("app/src/main/java/dev/amenhancer/module/hook/AppleMusicDualPaneTarget.kt"),
         ).firstOrNull(File::isFile)?.readText()
-            ?: error("DualPaneFeature.kt was not found from the unit-test working directory")
+            ?: error("AppleMusicDualPaneTarget.kt was not found from the unit-test working directory")
     }
     private val compactSource: String by lazy { source.replace(Regex("\\s+"), " ") }
